create pins with random adresses
