+++ css/style.css
@@ -1,7 +1,7 @@
 /* Общие стили. Каркас
    ========================================================================== */
 
-@font-face {
+/* @font-face {
   font-family: "Roboto";
   src: url("../fonts/Roboto-Bold.woff2") format("woff2"),
       url("../fonts/Roboto-Bold.woff") format("woff");
@@ -26,7 +26,7 @@
   font-weight: 500;
   font-style: normal;
   font-display: swap;
-}
+} */
 
 *,
 *::before,
