+++ js/main.js
@@ -1,51 +1,21 @@
 'use strict';
 
-var data = require('./data');
+var utils = require('./utils');
 
-var generateObjects = function () {
-  var objects = [];
-  for (var i = 0; i < data.titles.length; i++) {
-    var addressX = getRandomNumber(100, 1100);
-    var addressY = getRandomNumber(130, 630);
-    var price = getRandomNumber(500, 1300);
-    var indexType = getRandomNumber(0, 3);
-    var totalRooms = getRandomNumber(1, 4);
-    var totalGuests = getRandomNumber(1, 7);
-    var indexCheckin = getRandomNumber(0, 2);
-    var indexCheckout = getRandomNumber(0, 2);
-    var indexFeatures = getRandomNumber(0, 5);
-    var indexPhotos = getRandomNumber(0, 2);
-    var x = getRandomNumber(100, 1100);
-    var y = getRandomNumber(130, 630);
+var mapSection = document.querySelector('.map--faded');
+mapSection.classList.remove('map--faded');
+var mapPins = document.querySelector('.map__pins');
+var pinTemplate = document.querySelector('#pin').content;
+var pin = pinTemplate.querySelector('.map__pin');
 
-    var announcement = {
-      author: {
-        avatar: data.avatars[i],
-      },
-      offer: {
-        title: data.titles[i],
-        address: addressX + ', ' + addressY,
-        price: price,
-        type: data.types[indexType],
-        rooms: totalRooms,
-        guests: totalGuests,
-        checkin: data.time[indexCheckin],
-        checkout: data.time[indexCheckout],
-        features: data.features[indexFeatures],
-        description: data.descriptions[i],
-        photos: data.photos[indexPhotos],
-      },
-      location: {
-        x: x,
-        y: y
-      }
-    };
-    objects.push(announcement);
-  }
-  return objects;
-};
-function getRandomNumber(min, max) {
-  var random = min + Math.random() * (max + 1 - min);
-  return Math.floor(random);
-}
-console.log(generateObjects());
+var announcements = utils.generateObjects();
+
+announcements.forEach(function (announcement) {
+  var clonedPin = pin.cloneNode(true);
+  var image = clonedPin.querySelector('img');
+
+  image.src = announcement.author.avatar;
+  clonedPin.style.left = announcement.location.x + 'px';
+  clonedPin.style.top = announcement.location.y + 'px';
+  mapPins.appendChild(clonedPin);
+});
